+++ R/cdrs-package.R
@@ -1,0 +1,13 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom dplyr across
+#' @importFrom dplyr filter
+#' @importFrom dplyr mutate
+#' @importFrom dplyr select
+#' @importFrom magrittr %>%
+#' @importFrom rlang as_string
+#' @importFrom rlang enquos
+## usethis namespace: end
+NULL