+++ NEWS.md
@@ -1,3 +1,7 @@
-# cdrs (development version)
+# cdrs 0.1.0
 
-* Initial CRAN submission.
+Addition of `cdrs_subset` and `cdrs_design` which subset data according to requirements of `cdrs_design` and create a survey design object, respectively.
+
+# cdrs 0.0.0.9000
+
+Initial commit with `cdrs_read` and helpers.